requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
Pillow==10.4.0
pytest==8.3.2
pytest-django==4.8.0
